requirements.txt
Package            Version
------------------ ------------
beautifulsoup4     4.12.2
blinker            1.6.2
certifi            2023.5.7
charset-normalizer 3.1.0
click              8.1.3
coverage           7.2.7
dnspython          2.3.0
email-validator    2.0.0.post2
Flask              2.3.2
flask-behind-proxy 0.1.1
Flask-SQLAlchemy   3.0.5
Flask-WTF          1.1.1
form               0.0.1
gitdb              4.0.10
GitPython          3.1.31
greenlet           2.0.2
html5lib           1.1
idna               3.4
iniconfig          2.0.0
itsdangerous       2.1.2
Jinja2             3.1.2
lxml               4.9.3
MarkupSafe         2.1.3
numpy              1.24.3
packaging          23.1
pandas             2.0.2
Pillow             9.5.0
pip                23.1.2
pluggy             1.2.0
pycodestyle        2.10.0
pymssql            2.2.7
pypng              0.20220715.0
pytest             7.4.0
python-dateutil    2.8.2
python-form        0.2.3
pytz               2023.3
requests           2.31.0
setuptools         65.5.0
six                1.16.0
smmap              5.0.0
soupsieve          2.4.1
SQLAlchemy         2.0.16
typing_extensions  4.6.3
tzdata             2023.3
urllib3            2.0.2
webencodings       0.5.1
Werkzeug           2.3.6
WTForms            3.0.1
